requirements.txt
pdfplumber==0.10.3
spacy==3.7.2
transformers==4.37.2
torch==2.2.0
fpdf2==2.7.6
python-dotenv==1.0.1 